### 2018-11-28 #1106 @whedon
[PRE REVIEW]: Py-school-match: Matching algorithms to assign students to schools
**Submitting author:** @igarizio (<a href="http://orcid.org/0000-0002-8431-516X">Iacopo Garizio</a>)
**Repository:** <a href="https://github.com/igarizio/py-school-match" target ="_blank">https://github.com/igarizio/py-school-match</a>
**Version:** 0.2.0
**Reviewers:** jmhernan
