### 2018-11-28 #1106 @arfon
:wave: @igarizio, many thanks for your submission to JOSS.

This looks like a nice piece of software, but from first inspection it doesn't appear to be _research software_. In particular, quoting from our [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements):

> - The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

Could you please explain what the research challenge being addressed by this software is? Links to academic research on these topics would be helpful.
