### 2018-11-28 #1106 @igarizio
Hi @arfon, thank you for your quick response!
I will try to explain the main problem that this software addresses:

In many countries or states where schools cannot discriminate among students (this means, no tests, no interviews, no previous grades examination, etc.), there is a problem that always appears: How to assign students fairly and efficiently.
There are many proposed solutions. One could think that a simple "lottery" (random assignation) could solve the problem, but this overlooks the fact that there may be better alternatives where EVERY student is better off.
For example, imagine 2 students (S1 and S2) and 2 schools (A and B), with this preferences:
S1: A > B (S1 wants to go to A)
S2: B > A (S2 wants to go to B)

A simple "lottery" could result in S1 going to B, and S2 going to A (both students are assigned to their worst option).
Here is where the algorithms come into play: If we apply, for example, Stable Improvement Cycles (SIC) we would get, S1 going to A and S2 going to B. In this assignation both students are better off.
In the real world, the problems are not as straightforward as the one above, because of many restrictions, different preferences, among others. This means that there is no one-size-fits-all algorithm. The only way to select the best algorithm is to simulate the country's conditions and possible student's preferences and evaluate the resulting assignation.

The software helps researchers focus on the "simulation" and the "selection of the best algorithm" parts, as they do not have to develop from zero all the available algorithms in order to recommend an assignation mechanism.

While there are other libraries that implement matching algorithms, none of those (as far as I know) have the capability to implement quotas, restrictions, etc. (which is the most common case for the majority of countries/states [I actually do not know any country that does not have any sort of special restriction]).

Here are some publications that research that exact problem (and that could have been benefited with the package):
- Evaluation for Boston: https://economics.mit.edu/files/3021
- Evaluation for New York: https://seii.mit.edu/wp-content/uploads/2011/12/Paper-New-York-City-High-School-Math.pdf
 A more in-depth comparison/description: https://economics.mit.edu/files/3962
- I also made a comparison for the Chilean school system (where I recommend changing the current mechanism), if you want to see that one too, I will happily upload a copy.
- Also, If needed I can provide more technical information explaining the algorithms.

I am sorry for writing such a long text!
Thank you!
