### 2018-11-29 #1106 @igarizio
Thank you @arokem!
I will add those explanations as soon as possible.
