### 2018-11-30 9e5e5a1
Added JOSS badge
