+++ JOSS/paper.md
@@ -19,12 +19,17 @@ bibliography: paper.bib
 
 # Summary
 
-The problem of assigning students to schools in a fair way is a long-lasting problem
-with many proposed solutions. These solutions go from simple ``lotteries'' to more complex
-graph algorithms, all of which have strengths and weaknesses. Given the nature of the
-problem and the fact that in the real world there is an extra layer of
-requirements (quotas, special conditions, ranks of preference, etc.) it is vital to
-analyze and simulate every available option.
+In many countries where schools cannot discriminate among student applicants
+(no entrance exams, no interviews, no previous grades examination, etc.),
+there is an occurring problem: How to assign students fairly and efficiently.
+
+Along the years, there have been many proposed solutions to this problem. These
+vary from simple ``lotteries'' (random priority assignation) to more complex graph algorithms.
+Because each option has its own strengths and weaknesses, and given that in the
+real world there is an extra layer of requirements (quotas, special conditions, ranks 
+of preference, etc.) it is vital to analyze and simulate every available option.
+The correct selection and implementation of the algorithm can have serious 
+effects on efficiency and fairness.
 
 ``py-school-match`` is a Python library that implements multiple matching algorithms
 and allows users to simply specify the country's conditions  or requirements, and then
@@ -33,16 +38,16 @@ What makes ``py-school-match`` different from other libraries is that it is spec
 created to be used in the student-to-school assignation problem. Another distinctive
 characteristic is that it allows the use of quotas, priorities, capacities, among
 others, without much effort. These characteristics allow researchers to evaluate accurately
-the different algorithms, without the need of developing case-by-case solutions.
+the different algorithms without the need of developing case-by-case solutions.
 
 ``py-school-match`` implements the following algorithms:
 
-    - Top Trading Cycles (TTC)
-    - Deferred acceptance with multiple tie-breaking (DAMTB)
-    - Deferred acceptance with single tie-breaking (DASTB)
-    - Stable improvement cycles (SIC)
-    - Deferred Acceptance with multiple tie-breaking, plus stable cycles (MSIC)
-    - Deferred Acceptance with single tie-breaking, plus non-stable cycles (NSIC)
+- Top Trading Cycles (TTC)
+- Deferred acceptance with multiple tie-breaking (DAMTB)
+- Deferred acceptance with single tie-breaking (DASTB)
+- Stable improvement cycles (SIC)
+- Deferred Acceptance with multiple tie-breaking, plus stable cycles (MSIC)
+- Deferred Acceptance with single tie-breaking, plus non-stable cycles (NSIC)
     
 
 # References
