+++ JOSS/paper.bib
@@ -1,3 +1,81 @@
+@techreport{equitable_ttc,
+  title={The equitable top trading cycles mechanism for school choice},
+  author={Hakimov, Rustamdjan and Kesten, Onur},
+  year={2014},
+  institution={WZB discussion paper}
+}
+
+@article{school_mechanism_design,
+  title={School Choice: A Mechanism Design Approach},
+  volume={93},
+  DOI={10.1257/000282803322157061},
+  number={3},
+  journal={American Economic Review},
+  author={Abdulkadiroğlu, Atila and Sönmez, Tayfun},
+  year={2003},
+  pages={729–747}}
+
+@article{boston_match,
+  title={The Boston Public School Match},
+  volume={95},
+  number={2},
+  journal={American Economic Review},
+  author={Abdulkadiroğlu, Atila and others},
+  year={2005},
+  pages={368–371}}
+
+@article{new_york_match,
+  title={The New York City High School Match},
+  volume={95},
+  number={2},
+  journal={American Economic Review},
+  author={Abdulkadiroğlu, Atila and others},
+  year={2005}}
+
+@article{two_mechanismns,
+  title={A Tale of Two Mechanisms: Student Placement},
+  volume={84},
+  DOI={10.1006/jeth.1998.2469},
+  number={1},
+  journal={Journal of Economic Theory},
+  author={Balinski, Michel and Sönmez, Tayfun},
+  year={1999},
+  pages={73–94}}
+
+@article{stable_allocation,
+  title={The Stable Allocation (or Ordinal Transportation) Problem},
+  volume={27},
+  DOI={10.1287/moor.27.3.485.310},
+  number={3},
+  journal={Mathematics of Operations Research},
+  author={Baïou, Mourad and Balinski, Michel},
+  year={2002},
+  month={8},
+  pages={485–503}}
+
+@article{whats_the_matter_tie_breaking,
+  title={What's the Matter with Tie-Breaking? Improving Efficiency in School Choice},
+  volume={98},
+  DOI={10.1257/aer.98.3.669},
+  number={3},
+  journal={American Economic Review},
+  author={Erdil, Aytek and Ergin, Haluk},
+  year={2008},
+  pages={669–689}}
+
+@article{college_admision,
+  ISSN = {00029890, 19300972},
+  URL = {http://www.jstor.org/stable/2312726},
+  author = {Gale and Shapley},
+  journal = {The American Mathematical Monthly},
+  number = {1},
+  pages = {9-15},
+  publisher = {Mathematical Association of America},
+  title = {College Admissions and the Stability of Marriage},
+  volume = {69},
+  year = {1962}
+}
+
 @article{roth_1982,
   author = {Alvin E. Roth},
   title = {The Economics of Matching: Stability and Incentives},
