### 2018-12-10 9572e11
Added citations
