+++ JOSS/paper.bib
@@ -63,17 +63,17 @@
   year={2008},
   pages={669–689}}
 
-@article{college_admision,
-  ISSN = {00029890, 19300972},
-  URL = {http://www.jstor.org/stable/2312726},
-  author = {Gale and Shapley},
-  journal = {The American Mathematical Monthly},
-  number = {1},
-  pages = {9-15},
-  publisher = {Mathematical Association of America},
-  title = {College Admissions and the Stability of Marriage},
-  volume = {69},
-  year = {1962}
+@article{GS_1962,
+    title = {College Admissions and the Stability of Marriage},
+    ISSN = {00029890, 19300972},
+    URL = {http://www.jstor.org/stable/2312726},
+    author = {Roth and Shapley},
+    journal = {The American Mathematical Monthly},
+    number = {1},
+    pages = {9-15},
+    publisher = {Mathematical Association of America},
+    volume = {69},
+    year = {1962}
 }
 
 @article{roth_1982,
@@ -86,3 +86,15 @@
   year = {1982},
   doi = {10.1287/moor.7.4.617}
 }
+
+@article{strategy_vs_efficiency,
+  title={Strategy-proofness versus Efficiency in Matching with Indifferences: Redesigning the New York City High School Match},
+  author={Abdulkadiroğlu, Atila and others},
+  journal={The American Economic Review},
+  volume={99},
+  number={5},
+  pages={1954--1978},
+  year={2009},
+  publisher={American Economic Association},
+  DOI={10.3386/w14864},
+  }
