### 2018-12-10 10c30d8
Fixed citations
