+++ JOSS/paper.bib
@@ -20,7 +20,7 @@
   volume={95},
   number={2},
   journal={American Economic Review},
-  author={Abdulkadiroğlu, Atila and others},
+  author={Abdulkadiroğlu, Atila, Pathak, Parag and Roth, Alvin},
   year={2005},
   pages={368–371}}
 
@@ -29,7 +29,7 @@
   volume={95},
   number={2},
   journal={American Economic Review},
-  author={Abdulkadiroğlu, Atila and others},
+  author={Abdulkadiroğlu, Atila, Pathak, Parag and Roth, Alvin},
   year={2005}}
 
 @article{two_mechanismns,
@@ -89,7 +89,7 @@
 
 @article{strategy_vs_efficiency,
   title={Strategy-proofness versus Efficiency in Matching with Indifferences: Redesigning the New York City High School Match},
-  author={Abdulkadiroğlu, Atila and others},
+  author={Abdulkadiroğlu, Atila, Pathak, Parag and Roth, Alvin},
   journal={The American Economic Review},
   volume={99},
   number={5},
