### 2019-01-07 #1111 @jmhernan
@arokem I'm having trouble assessing this bullet:
> - [ ] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

The dependency for this software is `graph-tool` which is C++ library wrapped in python. The installation was not straight forward on my end (I could not get things to work on my conda python environment.  I was able to get it to work with homebrew after some stackoveflowing) Should more information be included or is a link to the `graph-tool` folks enough as is provided?
