### 2019-01-07 #1111 @arokem
I think that a link to this page: https://git.skewed.de/count0/graph-tool/wikis/installation-instructions would be helpful, considering installation can be a bit arcane.
