### 2019-01-08 #1111 @igarizio
@jmhernan Thank you for your comments! I will work on them as soon as posible.
