### 2019-01-11 18a92d1
Install instructions link added
