+++ docs/intro/test.rst
@@ -1,0 +1,15 @@
+.. _intro-test:
+
+=============
+Testing guide
+=============
+
+Testing py-school-match
+==========================
+
+To run the test suite, simply run::
+
+    python3 -m unittest discover
+
+
+
