### 2019-01-11 05ce3c8
Docstring wording change
