### 2019-01-11 c8ca96c
Comparison example added
