### 2019-01-11 5fc335c
Comparison example added
