+++ README.rst
@@ -92,7 +92,10 @@ Sample code
 
     # inspecting the obtained assignation
     for student in students:
-        print("Student {} was assigned to School {}".format(student.id, student.assigned_school.id))
+        if student.assigned_school is not None:
+            print("Student {} was assigned to School {}".format(student.id, student.assigned_school.id))
+        else:
+            print("Student {} was not assigned to any school".format(student.id))
 
 
 
