+++ README.rst
@@ -122,7 +122,7 @@ Or you can clone the repo and install it:
   cd py-school-match
   python setup.py install
 
-Remember to first install `graph-tool <https://graph-tool.skewed.de>`_ (`install instructions <https://git.skewed.de/count0/graph-tool/wikis/installation-instructions>`_).
+Remember to first install `graph-tool <https://graph-tool.skewed.de>`_ (`installation instructions <https://git.skewed.de/count0/graph-tool/wikis/installation-instructions>`_).
 
 Development
 -----------
