### 2019-01-11 #1111 @igarizio
@jmhernan I think I fixed all the comments. Here are the details:

> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I changed the wording of the paper (a little) and added a small introduction in the documentation page.
Answering your question, I think my software focuses only in researchers, because it is not fully optimized for each particular case available (something that practitioners may find really important).

> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I added a new example in the documentation page showing how to run and compare multiple algorithms.

> * [ ]  **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?

I added a section in the documentation showing how to run the test suite.

> * [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

I added the link to graph-tool's installation instructions in the README and in the docs.
