### 2019-01-27 #1111 @arokem
@igarizio : in the meanwhile, a couple of comments from me:

- The last sentence of the first paragraph should probably end with a question mark.
- I believe that "Along the years" (paragraph 2), should be "Over the years".
- "assignation" => "assignment" wherever it appears.
- "without the need of developing case-by-case solutions" => "without the need to develop case-by-case solutions".

I would prefer that the related literature section, that is currently a list of bullet points be integrated into your introductory paragraph, as references to specific points made there, or as extensions to that paragraph.
