### 2019-01-29 #1111 @igarizio
@arokem @jmhernan thank you for your feedback!
I think I have now fixed the issues in the paper (I have rephrased some parts of it). Please let me know if they sound good to you.
