### 2019-02-01 #1111 @arokem
@igarizio: looks good.

Could you please create an archive for the current state of the software (e.g., using [zenodo](https://zenodo.org/)) and post the doi that you get here?
