### 2019-02-04 #1111 @arokem
Congratulations @igarizio! Your paper is now ready to be accepted.

Thanks @jmhernan for the review!

Please stand by for the EIC or an Associate EIC to drop by and finalize this.

cc: @arfon, @danielskatz, @labarba, @kyleniemeyer
