### 2019-02-04 #1111 @arfon
One last thing @igarizio - could you please check to see if the references above under `MISSING DOIS` are indeed the correct DOIs for some of your references? If they are, then please add the DOI field to your bibtex.
