+++ JOSS/paper.bib
@@ -2,7 +2,8 @@
   title={The equitable top trading cycles mechanism for school choice},
   author={Hakimov, Rustamdjan and Kesten, Onur},
   year={2014},
-  institution={WZB discussion paper}
+  institution={WZB discussion paper},
+  DOI={10.1111/iere.12335}
 }
 
 @article{school_mechanism_design,
@@ -13,7 +14,8 @@
   journal={American Economic Review},
   author={Abdulkadiroğlu, Atila and Sönmez, Tayfun},
   year={2003},
-  pages={729–747}}
+  pages={729–747}
+}
 
 @article{boston_match,
   title={The Boston Public School Match},
@@ -22,7 +24,9 @@
   journal={American Economic Review},
   author={Abdulkadiroğlu, Atila, Pathak, Parag and Roth, Alvin},
   year={2005},
-  pages={368–371}}
+  pages={368–371},
+  DOI={10.1257/000282805774669637}
+}
 
 @article{new_york_match,
   title={The New York City High School Match},
@@ -30,7 +34,9 @@
   number={2},
   journal={American Economic Review},
   author={Abdulkadiroğlu, Atila, Pathak, Parag and Roth, Alvin},
-  year={2005}}
+  year={2005},
+  DOI={10.1257/000282805774670167}
+}
 
 @article{two_mechanismns,
   title={A Tale of Two Mechanisms: Student Placement},
@@ -40,7 +46,8 @@
   journal={Journal of Economic Theory},
   author={Balinski, Michel and Sönmez, Tayfun},
   year={1999},
-  pages={73–94}}
+  pages={73–94}
+}
 
 @article{stable_allocation,
   title={The Stable Allocation (or Ordinal Transportation) Problem},
@@ -51,7 +58,8 @@
   author={Baïou, Mourad and Balinski, Michel},
   year={2002},
   month={8},
-  pages={485–503}}
+  pages={485–503}
+}
 
 @article{whats_the_matter_tie_breaking,
   title={What's the Matter with Tie-Breaking? Improving Efficiency in School Choice},
@@ -61,19 +69,21 @@
   journal={American Economic Review},
   author={Erdil, Aytek and Ergin, Haluk},
   year={2008},
-  pages={669–689}}
+  pages={669–689}
+}
 
 @article{GS_1962,
-    title = {College Admissions and the Stability of Marriage},
-    ISSN = {00029890, 19300972},
-    URL = {http://www.jstor.org/stable/2312726},
-    author = {Roth and Shapley},
-    journal = {The American Mathematical Monthly},
-    number = {1},
-    pages = {9-15},
-    publisher = {Mathematical Association of America},
-    volume = {69},
-    year = {1962}
+  title = {College Admissions and the Stability of Marriage},
+  ISSN = {00029890, 19300972},
+  URL = {http://www.jstor.org/stable/2312726},
+  author = {Roth and Shapley},
+  journal = {The American Mathematical Monthly},
+  number = {1},
+  pages = {9-15},
+  publisher = {Mathematical Association of America},
+  volume = {69},
+  year = {1962},
+  DOI={10.2307/2312726}
 }
 
 @article{roth_1982,
@@ -97,4 +107,4 @@
   year={2009},
   publisher={American Economic Association},
   DOI={10.3386/w14864},
-  }
+}
