### 2019-02-04 #1111 @igarizio
@arfon I added the missing DOIs.
